Fix geo-tracking bugs from testing + extract NewVisitForm component
- Fix admin check: use level===4 instead of role==="admin" (end, detail, visit page)
- Swap ownership check before idempotency check in end-visit route
- Remove unreachable privacy filtering code in GET visit (403 already blocks)
- Extract NewVisitForm client component from visit/new page for server-side auth
- Add ended_at/inserted_at to visit history (SchoolTabs, VisitsTab)
- Show "Ended" badge (blue) distinct from "In Progress" (yellow)
- Remove unused allComplete variable (lint fix)
