+++ src/components/VisitsTab.tsx
@@ -7,6 +7,8 @@ interface Visit {
   id: number;
   visit_date: string;
   status: string;
+  inserted_at: string;
+  ended_at: string | null;
 }
 
 interface Props {
